requirements.txt
pandas
scikit-learn
torch
transformers
pyyaml
streamlit
plotly
tqdm
numpy
hf_xet
google.generativeai
psutil
pynvml
#torch torchvision --index-url https://download.pytorch.org/whl/cu126 # to use GPU if available